pyproject.toml
[project]
name = "imgtoascii"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
